refactor(listings): apply sourcery suggestions — split admin-listings-table into container + presentational, extract small UI components and utils, move/rename view file, remove top-level duplicates; backend: add applyToListing UoW helper, extract resolver helpers (buildPagedArgs, mapStateToStatus, toAdminListing), replace bracket access with dot-notation
