+++ apps/ui-sharethrift/src/components/layouts/home/account/admin-dashboard/components/admin-listings-table/admin-listings-table.tsx
@@ -1,10 +1,12 @@
-import { Input, Checkbox, Button, Image, Popconfirm, Tag } from 'antd';
+import { Button, Image, Popconfirm } from 'antd';
 import type { TableProps } from 'antd';
 import { SearchOutlined, FilterOutlined } from '@ant-design/icons';
-
-const { Search } = Input;
 import { Dashboard } from '@sthrift/ui-components';
 import type { MyListingData } from '../../../../my-listings/components/my-listings-dashboard.types.ts';
+import { TitleFilter } from './admin-listings-table.title-filter';
+import { StatusFilter } from './admin-listings-table.status-filter';
+import { StatusTag } from './admin-listings-table.status-tag';
+import { formatDate } from './admin-listings-table.utils';
 
 export interface AdminListingsTableProps {
   readonly data: ReadonlyArray<MyListingData>;
@@ -22,11 +24,6 @@ export interface AdminListingsTableProps {
   onAction: (action: string, listingId: string) => void;
 }
 
-const STATUS_OPTIONS = [
-  { label: 'Appealed', value: 'Appeal Requested' },
-  { label: 'Blocked', value: 'Blocked' },
-];
-
 export function AdminListingsTable({
   data,
   searchText,
@@ -42,133 +39,101 @@ export function AdminListingsTable({
   onPageChange,
   onAction,
 }: AdminListingsTableProps) {
-  // Compute sort orders outside of columns array to avoid nested ternaries
   let publishedAtSortOrder: 'ascend' | 'descend' | null = null;
-  if (sorter?.field === 'publishedAt') {
-    publishedAtSortOrder = sorter.order === 'ascend' ? 'ascend' : 'descend';
-  }
+  if (sorter?.field === 'publishedAt') publishedAtSortOrder = sorter.order === 'ascend' ? 'ascend' : 'descend';
 
   let reservationPeriodSortOrder: 'ascend' | 'descend' | null = null;
-  if (sorter?.field === 'reservationPeriod') {
-    reservationPeriodSortOrder = sorter.order === 'ascend' ? 'ascend' : 'descend';
-  }
+  if (sorter?.field === 'reservationPeriod') reservationPeriodSortOrder = sorter.order === 'ascend' ? 'ascend' : 'descend';
 
   const columns: TableProps<MyListingData>['columns'] = [
-      {
-        title: 'Listing',
-        dataIndex: 'title',
-        key: 'title',
-        width: 300,
-        filterDropdown: ({ setSelectedKeys, selectedKeys, confirm }) => (
-          <div style={{ padding: 8 }}>
-            <Search
-              placeholder="Search listings"
-              value={selectedKeys.length ? (selectedKeys[0] as string) : searchText}
-              onChange={(e) => {
-                setSelectedKeys(e.target.value ? [e.target.value] : []);
-              }}
-              onSearch={(value) => {
-                confirm();
-                onSearch(value);
-              }}
-              style={{ width: 250 }}
-              allowClear
-            />
-          </div>
-        ),
-        filterIcon: (filtered: boolean) => (
-          <SearchOutlined style={{ color: filtered ? '#1890ff' : undefined }} />
-        ),
-        render: (title: string, record: MyListingData) => (
-          <div style={{ display: 'flex', alignItems: 'center', gap: 12 }}>
-            <Image
-              src={record.image ?? ''}
-              alt={title}
-              width={72}
-              height={72}
-              style={{ objectFit: 'cover', borderRadius: 4 }}
-              preview={false}
-            />
-            <span>{title}</span>
-          </div>
-        ),
-      },
-      {
-        title: 'Published At',
-        dataIndex: 'publishedAt',
-        key: 'publishedAt',
-        sorter: true,
-        sortOrder: publishedAtSortOrder,
-        render: (date: string) => (date ? new Date(date).toLocaleDateString() : 'N/A'),
-      },
-      {
-        title: 'Reservation Period',
-        dataIndex: 'reservationPeriod',
-        key: 'reservationPeriod',
-        sorter: true,
-        sortOrder: reservationPeriodSortOrder,
-        render: (period: string) => period,
-      },
-      {
-        title: 'Status',
-        dataIndex: 'status',
-        key: 'status',
-        filterDropdown: () => (
-          <div style={{ padding: 16, width: 200 }}>
-            <div style={{ marginBottom: 8, fontWeight: 500 }}>Filter by Status</div>
-            <Checkbox.Group
-              options={STATUS_OPTIONS}
-              value={[...statusFilters]}
-              onChange={(checkedValues: Array<string | number>) => {
-                onStatusFilter(checkedValues.map(String));
-                confirm();
-              }}
-              style={{ display: 'flex', flexDirection: 'column', gap: 8 }}
-            />
-          </div>
-        ),
-        filterIcon: (filtered: boolean) => (
-          <FilterOutlined style={{ color: filtered ? '#1890ff' : undefined }} />
-        ),
-        render: (status: string) => {
-          const normalized = status === 'Appeal Requested' ? 'Appealed' : status;
-          const color = normalized === 'Appealed' ? 'gold' : 'purple';
-          return <Tag color={color}>{normalized}</Tag>;
-        },
-      },
-      {
-        title: 'Actions',
-        key: 'actions',
-        width: 240,
-        render: (_: unknown, record: MyListingData) => (
-          <div style={{ display: 'flex', gap: 8, justifyContent: 'flex-start' }}>
-            <Button type="link" size="small" onClick={() => onAction('view', record.id)}>
-              View
+    {
+      title: 'Listing',
+      dataIndex: 'title',
+      key: 'title',
+      width: 300,
+      filterDropdown: ({ setSelectedKeys, selectedKeys, confirm }) => (
+        <TitleFilter
+          searchText={searchText}
+          onSearch={onSearch}
+          setSelectedKeys={setSelectedKeys}
+          selectedKeys={selectedKeys}
+          confirm={confirm}
+        />
+      ),
+      filterIcon: (filtered: boolean) => <SearchOutlined style={{ color: filtered ? '#1890ff' : undefined }} />,
+      render: (title: string, record: MyListingData) => (
+        <div style={{ display: 'flex', alignItems: 'center', gap: 12 }}>
+          <Image
+            src={record.image ?? ''}
+            alt={title}
+            width={72}
+            height={72}
+            style={{ objectFit: 'cover', borderRadius: 4 }}
+            preview={false}
+          />
+          <span>{title}</span>
+        </div>
+      ),
+    },
+    {
+      title: 'Published At',
+      dataIndex: 'publishedAt',
+      key: 'publishedAt',
+      sorter: true,
+      sortOrder: publishedAtSortOrder,
+      render: (date: string) => formatDate(date),
+    },
+    {
+      title: 'Reservation Period',
+      dataIndex: 'reservationPeriod',
+      key: 'reservationPeriod',
+      sorter: true,
+      sortOrder: reservationPeriodSortOrder,
+      render: (period: string) => period,
+    },
+    {
+      title: 'Status',
+      dataIndex: 'status',
+      key: 'status',
+      filterDropdown: ({ confirm }) => (
+        <StatusFilter statusFilters={[...statusFilters]} onStatusFilter={onStatusFilter} confirm={confirm} />
+      ),
+      filterIcon: (filtered: boolean) => <FilterOutlined style={{ color: filtered ? '#1890ff' : undefined }} />,
+      render: (status: string) => <StatusTag status={status} />,
+    },
+    {
+      title: 'Actions',
+      key: 'actions',
+      width: 240,
+      render: (_: unknown, record: MyListingData) => (
+        <div style={{ display: 'flex', gap: 8, justifyContent: 'flex-start' }}>
+          <Button type="link" size="small" onClick={() => onAction('view', record.id)}>
+            View
+          </Button>
+          {(record.status === 'Blocked' || record.status === 'Appeal Requested') && (
+            <Button type="link" size="small" onClick={() => onAction('unblock', record.id)}>
+              Unblock
             </Button>
-            {(record.status === 'Blocked' || record.status === 'Appeal Requested') && (
-              <Button type="link" size="small" onClick={() => onAction('unblock', record.id)}>
-                Unblock
-              </Button>
-            )}
-            <Popconfirm
-              title="Remove this listing?"
-              description="This action cannot be undone."
-              onConfirm={() => onAction('remove', record.id)}
-              okText="Remove"
-              cancelText="Cancel"
-            >
-              <Button type="link" size="small" danger>
-                Remove
-              </Button>
-            </Popconfirm>
-          </div>
-        ),
-      },
-    ];
+          )}
+          <Popconfirm
+            title="Remove this listing?"
+            description="This action cannot be undone."
+            onConfirm={() => onAction('remove', record.id)}
+            okText="Remove"
+            cancelText="Cancel"
+          >
+            <Button type="link" size="small" danger>
+              Remove
+            </Button>
+          </Popconfirm>
+        </div>
+      ),
+    },
+  ];
 
   return (
     <Dashboard
-        data={[...data]}
+      data={[...data]}
       columns={columns}
       loading={loading}
       currentPage={currentPage}
