+++ packages/sthrift/application-services/src/contexts/listing/item/index.ts
@@ -45,6 +45,26 @@ export interface ItemListingApplicationService {
 export const ItemListing = (
 	dataSources: DataSources,
 ): ItemListingApplicationService => {
+	const uow = dataSources.domainDataSource.Listing.ItemListing.ItemListingUnitOfWork;
+
+	// helper for mutate–save in a scoped transaction
+	type ListingMutator = { requestDelete?: () => void; unblock?: () => void } & Record<string, unknown>;
+
+	const applyToListing = async (
+		id: string,
+		mutate: (listing: ListingMutator) => void,
+	): Promise<boolean> => {
+		let success = false;
+		await uow.withScopedTransaction(async (repo) => {
+			const listing = await repo.get(id);
+			// cast repository entity to the light mutator type for the callback
+			mutate(listing as unknown as ListingMutator);
+			await repo.save(listing);
+			success = true;
+		});
+		return success;
+	};
+
 	return {
 		create: create(dataSources),
 		queryById: queryById(dataSources),
@@ -76,28 +96,10 @@ export const ItemListing = (
 			return await dataSources.readonlyDataSource.Listing.ItemListing.ItemListingReadRepo.getPaged(args);
 		},
 		remove: async ({ id }) => {
-			let success = false;
-			await dataSources.domainDataSource.Listing.ItemListing.ItemListingUnitOfWork.withScopedTransaction(
-				async (repo) => {
-					const listing = await repo.get(id);
-					listing.requestDelete();
-					await repo.save(listing);
-					success = true;
-				},
-			);
-			return success;
+			return await applyToListing(id, (listing) => listing.requestDelete?.());
 		},
 		unblock: async ({ id }) => {
-			let success = false;
-			await dataSources.domainDataSource.Listing.ItemListing.ItemListingUnitOfWork.withScopedTransaction(
-				async (repo) => {
-					const listing = await repo.get(id);
-					listing.unblock();
-					await repo.save(listing);
-					success = true;
-				},
-			);
-			return success;
+			return await applyToListing(id, (listing) => listing.unblock?.());
 		},
 	};
 };
