+++ packages/sthrift/graphql/src/schema/types/listing/item-listing.resolvers.ts
@@ -1,5 +1,5 @@
 import type { GraphContext } from '../../../init/context.ts';
-import type { Domain } from '@sthrift/domain';
+// Domain types not needed in this file after refactor
 import { toGraphItem } from '../../../helpers/mapping.js';
 import type { CreateItemListingInput } from '../../builder/generated.js';
 
@@ -11,29 +11,86 @@ interface MyListingsArgs {
 	sorter?: { field: string; order: 'ascend' | 'descend' };
 }
 
+// Small helpers extracted from resolver logic to keep resolvers focused on orchestration
+function buildPagedArgs(
+	args: MyListingsArgs,
+	opts?: { useDefaultStatuses?: boolean },
+) {
+	const { page, pageSize, searchText, statusFilters, sorter } = args;
+		let effectiveStatuses: string[] | undefined;
+		if (statusFilters && statusFilters.length > 0) {
+			effectiveStatuses = statusFilters;
+		} else if (opts?.useDefaultStatuses) {
+			effectiveStatuses = ['Appeal Requested', 'Blocked'];
+		} else {
+			effectiveStatuses = undefined;
+		}
+
+	const pagedArgs: Partial<MyListingsArgs> & { page: number; pageSize: number; sharerId?: string } = {
+		page,
+		pageSize,
+	};
+	if (searchText) pagedArgs.searchText = searchText;
+	if (effectiveStatuses !== undefined) pagedArgs.statusFilters = effectiveStatuses;
+	if (sorter) pagedArgs.sorter = sorter;
+	return pagedArgs;
+}
+
+function mapStateToStatus(state?: string) {
+	return state?.toString?.().trim?.() || 'Unknown';
+}
+
+function toAdminListing(listing: Record<string, unknown>) {
+	const getIso = (v: unknown) => {
+		try {
+			return (v as { toISOString?: () => string })?.toISOString?.() ?? '';
+		} catch {
+			return '';
+		}
+	};
+
+	// Cast to any locally so we can use dot-notation while preserving the
+	// original dynamic shape coming from Mongoose documents.
+	const l = listing as any;
+	const start = getIso(l.sharingPeriodStart);
+	const end = getIso(l.sharingPeriodEnd);
+	const images = l.images as string[] | undefined;
+	const createdAt = getIso(l.createdAt);
+	const state = (l.state as string | undefined) ?? undefined;
+	return {
+		id: (l.id as string) ?? '',
+		title: (l.title as string) ?? '',
+		image: images && images.length > 0 ? images[0] : null,
+		publishedAt: createdAt || null,
+		reservationPeriod: `${start.slice(0, 10)} - ${end.slice(0, 10)}`,
+		status: mapStateToStatus(state),
+		pendingRequestsCount: 0,
+	};
+}
+
 const itemListingResolvers = {
 	Query: {
-		itemListings: async (
+		myListingsAll: async (
 			_parent: unknown,
-			_args: unknown,
+			args: MyListingsArgs,
 			context: GraphContext,
 		) => {
 			const currentUser = context.applicationServices.verifiedUser;
-			const user = currentUser?.verifiedJwt?.sub;
+			const sharerId = currentUser?.verifiedJwt?.sub;
+			// Build paged args and include sharerId for personal listings
+			const pagedArgs = buildPagedArgs(args);
+			if (sharerId !== undefined) pagedArgs.sharerId = sharerId;
+			const result = await context.applicationServices.Listing.ItemListing.queryPaged(pagedArgs);
 
-			let listings: Domain.Contexts.Listing.ItemListing.ItemListingEntityReference[];
-
-			if (user) {
-				listings =
-					await context.applicationServices.Listing.ItemListing.queryBySharer({
-						personalUser: user,
-					});
-			} else {
-				listings =
-					await context.applicationServices.Listing.ItemListing.queryAll({});
-			}
-
-			return listings.map(toGraphItem);
+			return {
+				items: result.items.map(
+					//biome-ignore lint/suspicious/noExplicitAny: Mongoose document type is dynamic
+					(listing: any) => toAdminListing(listing),
+				),
+				total: result.total,
+				page: result.page,
+				pageSize: result.pageSize,
+			};
 		},
 
 		itemListing: async (
@@ -54,106 +111,18 @@ const itemListingResolvers = {
 			return toGraphItem(listing);
 		},
 
-		myListingsAll: async (
-			_parent: unknown,
-			args: MyListingsArgs,
-			context: GraphContext,
-		) => {
-			const currentUser = context.applicationServices.verifiedUser;
-			const sharerId = currentUser?.verifiedJwt?.sub;
-			const { page, pageSize, searchText, statusFilters, sorter } = args;
-			const pagedArgs: {
-				page: number;
-				pageSize: number;
-				searchText?: string;
-				statusFilters?: string[];
-				sharerId?: string;
-				sorter?: { field: string; order: 'ascend' | 'descend' };
-			} = { page, pageSize };
-			if (searchText !== undefined) {
-				pagedArgs.searchText = searchText;
-			}
-			if (statusFilters !== undefined) {
-				pagedArgs.statusFilters = statusFilters;
-			}
-			if (sorter !== undefined) {
-				pagedArgs.sorter = sorter;
-			}
-			if (sharerId !== undefined) {
-				pagedArgs.sharerId = sharerId;
-			}
-			const result =
-				await context.applicationServices.Listing.ItemListing.queryPaged(
-					pagedArgs,
-				);
 
-			return {
-				items: result.items.map((
-					//biome-ignore lint/suspicious/noExplicitAny: Mongoose document type is dynamic
-					listing: any) => {
-					const sharingStart = listing.sharingPeriodStart.toISOString();
-					const sharingEnd = listing.sharingPeriodEnd.toISOString();
-					return {
-						id: listing.id,
-						title: listing.title,
-						image:
-						listing.images && listing.images.length > 0
-							? listing.images[0]
-							: null,
-					publishedAt: listing.createdAt.toISOString(),
-					reservationPeriod: `${sharingStart.slice(0, 10)} - ${sharingEnd.slice(0, 10)}`,
-					status: listing?.state || 'Unknown',
-						// TODO: integrate with reservation request domain context
-						pendingRequestsCount: 0,
-					};
-					}),
-					total: result.total,
-					page: result.page,
-					pageSize: result.pageSize,
-				};
-		},
 
-		adminListings: async (
-			_parent: unknown,
-			args: MyListingsArgs,
-			context: GraphContext,
-		) => {
+		adminListings: async (_parent: unknown, args: MyListingsArgs, context: GraphContext) => {
 			// TODO: SECURITY - Add admin role-based authorization check when admin role system is implemented
-			const { page, pageSize, searchText, statusFilters, sorter } = args;
-			const pagedArgs: {
-				page: number;
-				pageSize: number;
-				searchText?: string;
-				statusFilters?: string[];
-				sorter?: { field: string; order: 'ascend' | 'descend' };
-			} = { page, pageSize };
-			if (searchText !== undefined) pagedArgs.searchText = searchText;
-			if (statusFilters !== undefined) pagedArgs.statusFilters = statusFilters;
-			if (sorter !== undefined) pagedArgs.sorter = sorter;
-			
-			const result = await context.applicationServices.Listing.ItemListing.queryPaged(
-				pagedArgs,
-			);
+			const pagedArgs = buildPagedArgs(args, { useDefaultStatuses: true });
+			const result = await context.applicationServices.Listing.ItemListing.queryPaged(pagedArgs);
 
 			return {
-				items: result.items.map((
+				items: result.items.map(
 					//biome-ignore lint/suspicious/noExplicitAny: Mongoose document type is dynamic
-					listing: any) => {
-					const start = listing.sharingPeriodStart.toISOString();
-					const end = listing.sharingPeriodEnd.toISOString();
-					return {
-						id: listing.id,
-						title: listing.title,
-						image:
-							listing.images && listing.images.length > 0
-								? listing.images[0]
-								: null,
-						publishedAt: listing.createdAt?.toISOString?.() ?? null,
-						reservationPeriod: `${start.slice(0, 10)} - ${end.slice(0, 10)}`,
-						status: listing?.state || 'Unknown',
-						pendingRequestsCount: 0,
-					};
-				}),
+					(listing: any) => toAdminListing(listing),
+				),
 				total: result.total,
 				page: result.page,
 				pageSize: result.pageSize,
